// src/app/model/trdebtcollection.ts
export class trdebtcollection {
    debtCollectionNumber: number;
    companyContractNumber: string;
    customerNumber: string;
    paymentDate: string;
    fname: string;
    lname: string;
    accountNumber: string;
    idcard: string;
    debtCollectionStatus: string;
    documentationStatus: string;
    breakInterest: number;
    principle: number;
    interestNotRecognized: number;
    totalAmount: number;
    fee: number;
    JudgmentDay: string;
    expirationYear: string;
    groupStack: string;
    debtWriteOffDate: string;
    lawsuitExpired: string;
    outsourceCompany: string;
    yearOfPayment: string;
    monthOfPayment: string;
    legalNumber: string;
    moo: string;
    districts: string;
    amphures: string;
    province: string;
    postal: string;
    registrationNo1: string;
    deliveryDate1: string;
    note1: string;
    registrationNo2: string;
    deliveryDate2: string;
    note2: string;
    civilEffective: string;
    sentence: string;
    Legal: string;
    InvestigateDomicile: string;
    landOffice: string;
    InvestigateDomicileDate: string;
    domicilePropertyFound: string;
    foundProperty: string;
    adjoiningArea: string;
    InvestigationDayIn3: string;
    propertyNotes: string;
    undecidedCase: string;
    decidedCase: string;
    dateOfSeizure: string;
    dropAmount: number;
    refundDate: string;
    refund: string;
    dateOfAcceptingTheJob: string;
    sector: string;
    customerPaymentDate: string;
    paymentAmount: number;
    dateOfWithdrawingMoney: string;
    reimbursementAmount: number;
    bill: string;
    career: string;
    workName: string;
    workPhoneNumber: string;
    provincialCourt: string;
    socialSecurityCheckDate: string;
    workAddress1: string;
    workAddress2: string;
    summaryOfTheWork: string;
    typeOfWork: string;
    operationsOfTheFieldTeam: string;
    operatingProcedure: string;
    operator: string;
    teamWork: string;
    followUpTeam: string;
    note: string;
    trackDate: string;
    active: string;
    createdBy: string;
    createDate: Date; 
    updatedBy: string; 
    updateDate: Date; 
    status: string; 
    traceBy: string; 
};